v0.0.17 =======
To avoid issues with multiple versions of Tomes not being able to
recognize Tomes as Tomes, instanceof has been eradicated from Tomes.
Instead we now check the to see if the constructor's name has Tomes in it.
This is not perfect as Tomes will still fail to recognize Tomes from other
iframes and windows. It's a step in the right direction regardless.

Additionally, Tomes now has a readAll method to pull out the entire array
of diffs from the root Tome. It seems like the vast majority of use cases
always pull out all diffs, so this should improve performance there.