Add easy sync between independent tomes
Added 2 new methods:

sync(target, [all], [callback])
It syncs one or multiple tomes. target can be single tome or array of
tomes. all is a boolean and indicates if it syncs only the first
operation in stack or all of them. callback is an optional function
called after a sync with the diff object.

syncTo(target, [callback])
This function keeps source and target(s) automatically updated. callback
is an optional function called after a sync with the diff object.